put ContextProviderComponent in own file and export it
