+++ src/index.ts
@@ -6,3 +6,4 @@ global.HTMLElement = dom.window.HTMLElement;
 (global as any).customElements = dom.window.customElements;
 
 export {WebComponent} from './web-component';
+export {ContextProviderComponent} from './context-provider-component';
